requirements.txt
Flask
flask_sqlalchemy
PyMySQL
SQLAlchemy
python-dotenv